+++ Source/Shaders/SVGF/SVGFReproject_CS.hlsl
@@ -31,8 +31,8 @@ ConstantBuffer<DenoiseConstantBuffer>   PassCB                  : register(b0, s
 RWTexture2D<float4>                     PrevReprojMoments       : register(u0, space0);
 RWTexture2D<float4>                     CurrReprojMoments       : register(u1, space0);
 
-RWTexture2D<uint>                       PrevReprojHistoryLength : register(u2, space0);
-RWTexture2D<uint>                       CurrReprojHistoryLength : register(u3, space0);
+RWTexture2D<float>                      PrevReprojHistoryLength : register(u2, space0);
+RWTexture2D<float>                      CurrReprojHistoryLength : register(u3, space0);
 
 RWTexture2D<float4>                     PrevReprojDirect        : register(u4, space0);
 RWTexture2D<float4>                     CurrReprojDirect        : register(u5, space0);
